+++ crates/revm/src/context/context_precompiles.rs
@@ -1,26 +1,27 @@
 use crate::{
     precompile::{Precompile, PrecompileResult},
     primitives::{db::Database, Address, Bytes, HashMap},
-    EvmContext,
 };
 use core::ops::{Deref, DerefMut};
 use dyn_clone::DynClone;
 use revm_precompile::Precompiles;
 use std::{boxed::Box, sync::Arc};
 
+use super::InnerEvmContext;
+
 /// Precompile and its handlers.
-pub enum ContextPrecompile<DB: Database, EXTCTX> {
+pub enum ContextPrecompile<DB: Database> {
     /// Ordinary precompiles
     Ordinary(Precompile),
     /// Stateful precompile that is Arc over [`ContextStatefulPrecompile`] trait.
     /// It takes a reference to input, gas limit and Context.
-    ContextStateful(ContextStatefulPrecompileArc<EvmContext<DB>, EXTCTX>),
+    ContextStateful(ContextStatefulPrecompileArc<DB>),
     /// Mutable stateful precompile that is Box over [`ContextStatefulPrecompileMut`] trait.
     /// It takes a reference to input, gas limit and context.
-    ContextStatefulMut(ContextStatefulPrecompileBox<EvmContext<DB>, EXTCTX>),
+    ContextStatefulMut(ContextStatefulPrecompileBox<DB>),
 }
 
-impl<DB: Database, EXTCTX> Clone for ContextPrecompile<DB, EXTCTX> {
+impl<DB: Database> Clone for ContextPrecompile<DB> {
     fn clone(&self) -> Self {
         match self {
             Self::Ordinary(arg0) => Self::Ordinary(arg0.clone()),
@@ -31,11 +32,11 @@ impl<DB: Database, EXTCTX> Clone for ContextPrecompile<DB, EXTCTX> {
 }
 
 #[derive(Clone)]
-pub struct ContextPrecompiles<DB: Database, EXTCTX> {
-    inner: HashMap<Address, ContextPrecompile<DB, EXTCTX>>,
+pub struct ContextPrecompiles<DB: Database> {
+    inner: HashMap<Address, ContextPrecompile<DB>>,
 }
 
-impl<DB: Database, EXTCTX> ContextPrecompiles<DB, EXTCTX> {
+impl<DB: Database> ContextPrecompiles<DB> {
     /// Returns precompiles addresses.
     #[inline]
     pub fn addresses(&self) -> impl Iterator<Item = &Address> {
@@ -48,7 +49,7 @@ impl<DB: Database, EXTCTX> ContextPrecompiles<DB, EXTCTX> {
     #[inline]
     pub fn extend(
         &mut self,
-        other: impl IntoIterator<Item = impl Into<(Address, ContextPrecompile<DB, EXTCTX>)>>,
+        other: impl IntoIterator<Item = impl Into<(Address, ContextPrecompile<DB>)>>,
     ) {
         self.inner.extend(other.into_iter().map(Into::into));
     }
@@ -61,22 +62,19 @@ impl<DB: Database, EXTCTX> ContextPrecompiles<DB, EXTCTX> {
         addess: Address,
         bytes: &Bytes,
         gas_price: u64,
-        evmctx: &mut EvmContext<DB>,
-        extctx: &mut EXTCTX,
+        evmctx: &mut InnerEvmContext<DB>,
     ) -> Option<PrecompileResult> {
         let precompile = self.inner.get_mut(&addess)?;
 
         match precompile {
             ContextPrecompile::Ordinary(p) => Some(p.call(bytes, gas_price, &evmctx.env)),
-            ContextPrecompile::ContextStatefulMut(p) => {
-                Some(p.call_mut(bytes, gas_price, evmctx, extctx))
-            }
-            ContextPrecompile::ContextStateful(p) => Some(p.call(bytes, gas_price, evmctx, extctx)),
+            ContextPrecompile::ContextStatefulMut(p) => Some(p.call_mut(bytes, gas_price, evmctx)),
+            ContextPrecompile::ContextStateful(p) => Some(p.call(bytes, gas_price, evmctx)),
         }
     }
 }
 
-impl<DB: Database, EXTCTX> Default for ContextPrecompiles<DB, EXTCTX> {
+impl<DB: Database> Default for ContextPrecompiles<DB> {
     fn default() -> Self {
         Self {
             inner: Default::default(),
@@ -84,15 +82,15 @@ impl<DB: Database, EXTCTX> Default for ContextPrecompiles<DB, EXTCTX> {
     }
 }
 
-impl<DB: Database, EXTCTX> Deref for ContextPrecompiles<DB, EXTCTX> {
-    type Target = HashMap<Address, ContextPrecompile<DB, EXTCTX>>;
+impl<DB: Database> Deref for ContextPrecompiles<DB> {
+    type Target = HashMap<Address, ContextPrecompile<DB>>;
 
     fn deref(&self) -> &Self::Target {
         &self.inner
     }
 }
 
-impl<DB: Database, EXTCTX> DerefMut for ContextPrecompiles<DB, EXTCTX> {
+impl<DB: Database> DerefMut for ContextPrecompiles<DB> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.inner
     }
@@ -100,45 +98,41 @@ impl<DB: Database, EXTCTX> DerefMut for ContextPrecompiles<DB, EXTCTX> {
 
 /// Context aware stateful precompile trait. It is used to create
 /// a arc precompile in [`ContextPrecompile`].
-pub trait ContextStatefulPrecompile<EVMCTX, EXTCTX>: Sync + Send {
+pub trait ContextStatefulPrecompile<DB: Database>: Sync + Send {
     fn call(
         &self,
         bytes: &Bytes,
         gas_price: u64,
-        evmctx: &mut EVMCTX,
-        extctx: &mut EXTCTX,
+        evmctx: &mut InnerEvmContext<DB>,
     ) -> PrecompileResult;
 }
 
 /// Context aware mutable stateful precompile trait. It is used to create
 /// a boxed precompile in [`ContextPrecompile`].
-pub trait ContextStatefulPrecompileMut<EVMCTX, EXTCTX>: DynClone + Send + Sync {
+pub trait ContextStatefulPrecompileMut<DB: Database>: DynClone + Send + Sync {
     fn call_mut(
         &mut self,
         bytes: &Bytes,
         gas_price: u64,
-        evmctx: &mut EVMCTX,
-        extctx: &mut EXTCTX,
+        evmctx: &mut InnerEvmContext<DB>,
     ) -> PrecompileResult;
 }
 
-dyn_clone::clone_trait_object!(<EVMCTX, EXTCTX> ContextStatefulPrecompileMut<EVMCTX, EXTCTX>);
+dyn_clone::clone_trait_object!(<DB> ContextStatefulPrecompileMut<DB>);
 
 /// Arc over context stateful precompile.
-pub type ContextStatefulPrecompileArc<EVMCTX, EXTCTX> =
-    Arc<dyn ContextStatefulPrecompile<EVMCTX, EXTCTX>>;
+pub type ContextStatefulPrecompileArc<DB> = Arc<dyn ContextStatefulPrecompile<DB>>;
 
 /// Box over context mutable stateful precompile
-pub type ContextStatefulPrecompileBox<EVMCTX, EXTCTX> =
-    Box<dyn ContextStatefulPrecompileMut<EVMCTX, EXTCTX>>;
+pub type ContextStatefulPrecompileBox<DB> = Box<dyn ContextStatefulPrecompileMut<DB>>;
 
-impl<DB: Database, EXTCTX> From<Precompile> for ContextPrecompile<DB, EXTCTX> {
+impl<DB: Database> From<Precompile> for ContextPrecompile<DB> {
     fn from(p: Precompile) -> Self {
         ContextPrecompile::Ordinary(p)
     }
 }
 
-impl<DB: Database, EXTCTX> From<Precompiles> for ContextPrecompiles<DB, EXTCTX> {
+impl<DB: Database> From<Precompiles> for ContextPrecompiles<DB> {
     fn from(p: Precompiles) -> Self {
         ContextPrecompiles {
             inner: p.inner.into_iter().map(|(k, v)| (k, v.into())).collect(),
@@ -146,7 +140,7 @@ impl<DB: Database, EXTCTX> From<Precompiles> for ContextPrecompiles<DB, EXTCTX> 
     }
 }
 
-impl<DB: Database, EXTCTX> From<&Precompiles> for ContextPrecompiles<DB, EXTCTX> {
+impl<DB: Database> From<&Precompiles> for ContextPrecompiles<DB> {
     fn from(p: &Precompiles) -> Self {
         ContextPrecompiles {
             inner: p